+++ src/pages/GistGitHub.ts
@@ -4,6 +4,12 @@ import { getIconForFile, getIconUrl } from '../utils/Icons';
 const showGistIcons = async () => {
   const fileInfos = document.querySelectorAll('.file-info');
   for (let i = 0; i < fileInfos.length; i++) {
+    /**
+     * [DIV:
+     *  [SPAN: [SVG: icon]],
+     *  [A: [STRONG: name]]
+     * ]
+     */
     const fileInfo = fileInfos[i] as HTMLDivElement;
     const gistName = (fileInfo.lastElementChild!.firstElementChild as HTMLSpanElement).innerText;
     const iconPath = getIconForFile(gistName);