+++ src/pages/SourceForge.ts
@@ -1,0 +1,69 @@
+import { getIconForFile, getIconForFolder, getIconUrl } from '../utils/Icons';
+import { isSourceForgeFiles } from '../utils/PageDetect';
+
+const QUERY_SOURCEFORGE_ITEMS = '#files_list>tbody>tr';
+
+function showIconsForFiles() {
+  const items = document.querySelectorAll(QUERY_SOURCEFORGE_ITEMS);
+  for (let i = 0; i < items.length; i++) {
+    const item = items[i];
+
+    const newIconEl = document.createElement('img');
+    newIconEl.setAttribute('class', 'vscode-icon sf-icon');
+    const iconAndNameEl = item.firstElementChild as HTMLTableHeaderCellElement;
+
+    const isFolder = item.className.includes('folder');
+    if (isFolder) {
+      /**
+       * [TR:
+       *  [TH: [SPAN: icon]],
+       *  [TD: [ABBR: date]],
+       *  [TD: size],
+       *  [TD: [A: [IMG: chart]]],
+       *  [TD: [SPAN: info]]
+       * ]
+       */
+
+      const iconEl = iconAndNameEl.firstElementChild as HTMLSpanElement;
+      const nameEl = iconAndNameEl.lastElementChild as HTMLAnchorElement;
+      const name = nameEl.innerText.toLowerCase();
+      const iconPath = getIconForFolder(name);
+      newIconEl.setAttribute('src', getIconUrl(iconPath));
+      iconAndNameEl.replaceChild(newIconEl, iconEl);
+    } else {
+      /**
+       * [TR:
+       *  [TH: [[SPAN: icon], [A: name]]],
+       *  [TD: [ABBR: date]],
+       *  [TD: size],
+       *  [TD: [A: [IMG: chart]]],
+       *  [TD: [SPAN: info]]
+       * ]
+       */
+
+      const nameEl = iconAndNameEl.firstElementChild as HTMLAnchorElement;
+      const name = nameEl.innerText.toLowerCase();
+      const iconPath = getIconForFile(name);
+      newIconEl.setAttribute('src', getIconUrl(iconPath));
+      iconAndNameEl.insertBefore(newIconEl, nameEl);
+    }
+  }
+}
+
+const domLoaded = new Promise(resolve => {
+  if (document.readyState === 'loading') {
+    document.addEventListener('DOMContentLoaded', resolve);
+  } else {
+    resolve();
+  }
+});
+
+function update(e?: any) {
+  if (isSourceForgeFiles) {
+    showIconsForFiles();
+  }
+}
+
+export function initSourceForge() {
+  update();
+}