+++ src/utils/PageDetect.ts
@@ -2,29 +2,11 @@
 
 import * as select from 'select-dom';
 
-export const isDashboard = () => location.pathname === '/' || /^(\/orgs\/[^/]+)?\/dashboard/.test(location.pathname);
-
 export const isTrending = () => location.pathname.startsWith('/trending');
 
 export const getRepoPath = () => location.pathname.replace(/^\/[^/]+\/[^/]+/, '');
 
-export const getRepoURL = () =>
-  location.pathname
-    .slice(1)
-    .split('/', 2)
-    .join('/');
-
 export const isRepoTree = () => isRepo() && /\/tree\//.test(getRepoPath());
-
-export const isIssueSearch = () => location.pathname.startsWith('/issues');
-
-export const isIssueList = () => isRepo() && /^\/issues\/?$/.test(getRepoPath());
-
-export const isIssue = () => isRepo() && /^\/issues\/\d+/.test(getRepoPath());
-
-export const isPRSearch = () => location.pathname.startsWith('/pulls');
-
-export const isPRList = () => isRepo() && /^\/pulls\/?$/.test(getRepoPath());
 
 export const isPR = () => isRepo() && /^\/pull\/\d+/.test(getRepoPath());
 
@@ -32,31 +14,15 @@ export const isPRFiles = () => isRepo() && /^\/pull\/\d+\/files/.test(getRepoPat
 
 export const isPRCommit = () => isRepo() && /^\/pull\/\d+\/commits\/[0-9a-f]{5,40}/.test(getRepoPath());
 
-export const isMilestoneList = () => isRepo() && /^\/milestones\/?$/.test(getRepoPath());
-
-export const isMilestone = () => isRepo() && /^\/milestone\/\d+/.test(getRepoPath());
-
-export const isLabelList = () => isRepo() && /^\/labels\/?(((?=\?).*)|$)/.test(getRepoPath());
-
-export const isLabel = () => isRepo() && /^\/labels\/\w+/.test(getRepoPath());
-
-export const isCommitList = () => isRepo() && /^\/commits\//.test(getRepoPath());
-
 export const isSingleCommit = () => isRepo() && /^\/commit\/[0-9a-f]{5,40}/.test(getRepoPath());
 
 export const isCompare = () => isRepo() && /^\/compare/.test(getRepoPath());
-
-export const isQuickPR = () => isCompare() && /[?&]quick_pull=1(&|$)/.test(location.search);
-
-export const hasCode = () => isRepo() && select.exists('.highlight');
 
 export const hasDiff = () =>
   isRepo() &&
   (isSingleCommit() || isPRCommit() || isPRFiles() || isCompare() || (isPR() && select.exists('.diff-table')));
 
 export const isReleases = () => isRepo() && /^\/(releases|tags)/.test(getRepoPath());
-
-export const isBlame = () => isRepo() && /^\/blame\//.test(getRepoPath());
 
 export const isNotifications = () => location.pathname.startsWith('/notifications');
 
@@ -111,3 +77,8 @@ export const isPastebinUserList = () =>
   location.href.indexOf('pastebin.com/u/') > 0 && select.exists('table.maintable');
 
 export const isPasteOpen = () => select.exists('#code_frame2');
+
+/**
+ * SourceForge related detections
+ */
+export const isSourceForgeFiles = () => select.exists('#files_list');